Refactor TeamMember, OrgMember and Volunteer into subpackages of Person
