+++ src/main/java/loopin/projectbook/logic/parser/AddOrgMemberCommandParser.java
@@ -15,8 +15,8 @@ import loopin.projectbook.logic.commands.AddOrgMemberCommand;
 import loopin.projectbook.logic.parser.exceptions.ParseException;
 import loopin.projectbook.model.person.Email;
 import loopin.projectbook.model.person.Name;
-import loopin.projectbook.model.person.OrgMember;
-import loopin.projectbook.model.person.Organisation;
+import loopin.projectbook.model.person.orgmember.OrgMember;
+import loopin.projectbook.model.person.orgmember.Organisation;
 import loopin.projectbook.model.person.Phone;
 import loopin.projectbook.model.person.Telegram;
 import loopin.projectbook.model.tag.Tag;
