+++ src/main/java/loopin/projectbook/logic/parser/AddVolunteerCommandParser.java
@@ -16,7 +16,7 @@ import loopin.projectbook.model.person.Email;
 import loopin.projectbook.model.person.Name;
 import loopin.projectbook.model.person.Phone;
 import loopin.projectbook.model.person.Telegram;
-import loopin.projectbook.model.person.Volunteer;
+import loopin.projectbook.model.person.volunteer.Volunteer;
 import loopin.projectbook.model.tag.Tag;
 
 /**
