+++ src/main/java/loopin/projectbook/logic/parser/ParserUtil.java
@@ -11,13 +11,13 @@ import loopin.projectbook.commons.util.StringUtil;
 import loopin.projectbook.logic.parser.exceptions.ParseException;
 import loopin.projectbook.model.person.Email;
 import loopin.projectbook.model.person.Name;
-import loopin.projectbook.model.person.Organisation;
+import loopin.projectbook.model.person.orgmember.Organisation;
 import loopin.projectbook.model.person.Phone;
 import loopin.projectbook.model.person.Telegram;
 import loopin.projectbook.model.project.Description;
 import loopin.projectbook.model.project.ProjectName;
 import loopin.projectbook.model.tag.Tag;
-import loopin.projectbook.model.teammember.Committee;
+import loopin.projectbook.model.person.teammember.Committee;
 
 /**
  * Contains utility methods used for parsing strings in the various *Parser classes.
