+++ src/main/java/loopin/projectbook/model/person/volunteer/Volunteer.java
@@ -1,7 +1,12 @@
-package loopin.projectbook.model.person;
+package loopin.projectbook.model.person.volunteer;
 
 import java.util.Set;
 
+import loopin.projectbook.model.person.Email;
+import loopin.projectbook.model.person.Name;
+import loopin.projectbook.model.person.Person;
+import loopin.projectbook.model.person.Phone;
+import loopin.projectbook.model.person.Telegram;
 import loopin.projectbook.model.tag.Tag;
 
 /**
